pyproject.toml
[project]
name = "vamp-interface"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28",
    "omegaconf>=2.3",
    "numpy>=2.2",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.33.1",
    "scikit-learn>=1.6",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "transformers>=5.5.0",
    "pacmap>=0.9.1",
    "embedding-atlas>=0.20.0",
    "pandas>=3.0.2",
    "hdbscan>=0.8.42",
    "timm>=1.0.26",
    "ultralytics>=8.4.40",
    "gdown>=6.0.0",
    "dlib>=20.0.1",
    "insightface>=0.7.3",
    "onnxruntime-gpu>=1.24.4",
    "einops>=0.8.2",
    "sentencepiece>=0.2.1",
    "mediapipe>=0.10.33",
    "open-clip-torch>=3.3.0",
    "diffusers>=0.37.1",
    "peft>=0.19.1",
    "bitsandbytes>=0.49.2",
    "accelerate>=1.13.0",
    "protobuf>=7.34.1",
    "onnx2torch>=1.5.15",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

